Move AppInfo.kt to appmanager/mobile
