+++ appmanager/mobile/src/androidMain/res/values/plurals.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <plurals name="appinfo_requested_permissions_count">
+        <item quantity="one">%s permission requested</item>
+        <item quantity="other">%s permissions requested</item>
+    </plurals>
+</resources>
