+++ appmanager/mobile/src/androidMain/res/values/strings.xml
@@ -1,6 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="appmanager_section_user_apps">User Apps</string>
-    <string name="appmanager_section_disabled_apps">Disabled Apps</string>
-    <string name="appmanager_section_system_apps">System Apps</string>
+    <string name="appmanager_section_user">User Apps</string>
+    <string name="appmanager_section_disabled">Disabled Apps</string>
+    <string name="appmanager_section_system">System Apps</string>
+    <string name="appinfo_version">Version %s</string>
+    <string name="appinfo_first_installed">First installed: %s</string>
+    <string name="appinfo_last_updated">Last updated: %s</string>
+    <string name="appinfo_open">Open</string>
+    <string name="appinfo_uninstall">Uninstall</string>
 </resources>
