+++ mobile/src/main/kotlin/com/boswelja/smartwatchextensions/devicemanagement/ui/info/WatchInfoScreen.kt
@@ -36,7 +36,7 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.boswelja.smartwatchextensions.R
-import com.boswelja.smartwatchextensions.common.ui.BigButton
+import com.boswelja.smartwatchextensions.appmanager.ui.BigButton
 import com.boswelja.smartwatchextensions.devicemanagement.Capability
 import com.boswelja.smartwatchextensions.devicemanagement.labelRes
 import com.boswelja.watchconnection.common.Watch
