+++ mobile/src/main/res/values/strings.xml
@@ -51,11 +51,6 @@
 
     <!-- App Manager strings -->
     <string name="main_app_manager_title">App Manager</string>
-    <string name="app_info_version_prefix">Version %s</string>
-    <string name="app_info_first_installed_prefix">First installed: %s</string>
-    <string name="app_info_last_updated_prefix">Last updated: %s</string>
-    <string name="app_info_open_button">Open</string>
-    <string name="app_info_uninstall_button">Uninstall</string>
     <string name="app_manager_watch_disconnected">We can\'t reach your watch right now. You may still view your apps, but you cannot make any changes</string>
     <string name="watch_connection_dismiss">Dismiss</string>
